validation, rating as float

+++ app/src/main/java/cz/upce/personallibrary/model/BookRating.kt
@@ -2,9 +2,9 @@ package cz.upce.personallibrary.model
 
 import java.io.Serializable
 
-data class BookRating(val value: Int) : Comparable<BookRating>, Serializable {
+data class BookRating(val value: Float) : Comparable<BookRating>, Serializable {
     init {
-        if (value < 1 || value > 5) {
+        if (value < 1.0f || value > 5.0f) {
             throw IllegalArgumentException("Rating out of range")
         }
     }
